+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 12.8
 - Fix PHP warning undefined array key when Meal Item Short Name was edited in MenuReports.php
 - Fix PHP8.5 deprecated using null as an array offset, use an empty string instead in MenuReports.php & ActivityReport.php
 - Fix PHP warning undefined array key when switching Meals in ServeMenus.php
+- Fix SQL error when no Quarters found (Course widget) in Course.php
 
 Changes in 12.7.2
 -----------------